+++ packages/legend-application-studio/src/components/editor/command-center/ProjectSearchCommand.tsx
@@ -148,6 +148,7 @@ export const ProjectSearchCommand = observer(() => {
             darkMode={true}
             formatOptionLabel={getPackageableElementOptionFormatter({
               darkMode: true,
+              pureModel: editorStore.graphManagerState.graph,
             })}
           />
         </div>