+++ packages/legend-application-studio/src/components/editor/side-bar/CreateNewElementModal.tsx
@@ -233,6 +233,7 @@ const NewPureModelConnectionDriverEditor = observer(
             darkMode={true}
             formatOptionLabel={getPackageableElementOptionFormatter({
               darkMode: true,
+              pureModel: editorStore.graphManagerState.graph,
             })}
           />
         </div>