+++ packages/legend-query-builder/src/components/workflows/MappingQueryBuilder.tsx
@@ -151,6 +151,7 @@ const MappingQueryBuilderSetupPanelContent = observer(
                 filterOption={mappingFilterOption}
                 formatOptionLabel={getPackageableElementOptionFormatter({
                   darkMode: !applicationStore.TEMPORARY__isLightThemeEnabled,
+                  pureModel: queryBuilderState.graphManagerState.graph,
                 })}
               />
             </div>
@@ -173,6 +174,7 @@ const MappingQueryBuilderSetupPanelContent = observer(
                 filterOption={runtimeFilterOption}
                 formatOptionLabel={getRuntimeOptionFormatter({
                   darkMode: !applicationStore.TEMPORARY__isLightThemeEnabled,
+                  pureModel: queryBuilderState.graphManagerState.graph,
                 })}
               />
             </div>