+++ packages/legend-query-builder/src/index.ts
@@ -45,6 +45,9 @@ export type { MappingRuntimeCompatibilityAnalysisResult } from './graphManager/a
 export * from './stores/ServiceInfo.js';
 export * from './components/ServiceQuerySetupUtils.js';
 export * from './components/QuerySetupUtils.js';
+export * from './components/QueryBuilderTextEditor.js';
+
+export { TEST__setUpQueryBuilder } from './components/QueryBuilderComponentTestUtils.js';
 
 // ------------------------------------------- Shared components -------------------------------------------
 
@@ -56,3 +59,4 @@ export { LambdaEditorState } from './stores/shared/LambdaEditorState.js';
 export * from './stores/shared/LambdaParameterState.js';
 export * from './stores/shared/ValueSpecificationModifierHelper.js';
 export * from './stores/shared/ValueSpecificationEditorHelper.js';
+export { QueryBuilderTextEditorMode } from './stores/QueryBuilderTextEditorState.js';