+++ packages/legend-server-sdlc/src/SDLCServerClient.ts
@@ -161,8 +161,15 @@ export class SDLCServerClient extends AbstractServerClient {
   get currentUserUrl(): string {
     return `${this.baseUrl}/currentUser`;
   }
+
+  get usersUrl(): string {
+    return `${this.baseUrl}/users`;
+  }
+
   getCurrentUser = (): Promise<PlainObject<User>> =>
     this.get(this.currentUserUrl);
+  getUsers = (name: string): Promise<PlainObject<User>[]> =>
+    this.get(this.usersUrl, undefined, undefined, { search: name });
 
   // ------------------------------------------- Project -------------------------------------------